+++ src/store/accountsSlice.js
@@ -7,10 +7,20 @@ import {
   removeTransaction,
 } from './transactionsSlice';
 
+import checkingSchema from '@/schemas/checking.json';
+import savingsSchema from '@/schemas/savings.json';
+import creditCardSchema from '@/schemas/creditCard.json';
+
 export const AccountType = Object.freeze({
   SAVINGS: 'Savings',
   CHECKING: 'Checking',
   CREDIT_CARD: 'Credit Card',
+});
+
+export const AccountSchema = Object.freeze({
+  [AccountType.SAVINGS]: savingsSchema,
+  [AccountType.CHECKING]: checkingSchema,
+  [AccountType.CREDIT_CARD]: creditCardSchema,
 });
 
 const initialState = {
@@ -82,6 +92,7 @@ const { addAccount, removeAccount, updateAccount } = accountsSlice.actions;
 export { updateAccount };
 
 const generateAccountObject = (id, name, type, transactions) => ({
+  version: AccountSchema[type].version,
   id,
   name,
   type,
@@ -109,6 +120,14 @@ const loadAccountFromFile = async () => {
 
 export const loadAccountAsync = () => async (dispatch) => {
   const data = await loadAccountFromFile();
+  if (!data) return;
+  if (data.version !== AccountSchema[data.type].version) {
+    console.log(
+      'Version mismatch',
+      AccountSchema[data.type].version,
+      data.version
+    );
+  }
   const account = generateAccountObject(
     data.id,
     data.name,
@@ -124,6 +143,17 @@ export const editAccountName = (id, name) => (dispatch) => {
 
 export const editAccountType = (id, type) => (dispatch) => {
   dispatch(updateAccount({ id, type }));
+};
+
+export const saveAccount = (account, filename) => {
+  const saveString = JSON.stringify(account, null, 2);
+  const saveBlob = new Blob([saveString]);
+  const url = URL.createObjectURL(saveBlob);
+  const link = document.createElement('a');
+  link.download = filename;
+  link.href = url;
+  link.click();
+  URL.revokeObjectURL(url);
 };
 
 export const selectAccountById = (id) => (state) =>
